abstracting users table
